manual integration Integrate 113308 from 3 to main: Port Qt 3.3 to MinGW.
Fix some of the warnings from the gcc compiler. Some of them
may be hiding real errors. Please review.

kernel/qapplication_win.cpp:268:
	warning: `UINT appUniqueID' defined but not used
=> This one is just cosmetic but trivial to fix.

kernel/qdnd_win.cpp:475:
	warning: enumeration value `DragLink' not handled
	in switch
kernel/qdnd_win.cpp:460:
	warning: `DWORD allowed_effects' might be used
	uninitialized in this function
=> I'm not 100% sure how to fix this, but what happens
   indeed if `mode' is `DragLink'?

kernel/qfont_win.cpp:57:
	warning: `int max_font_count' defined but not used
=> This one is just cosmetic but trivial to fix.

kernel/qfontdatabase_win.cpp:35:
	warning: `stock_sysfont' initialized and declared `extern'
=> *Please review* because something's seriously wrong here:
   $ grep stock_sysfont src/kernel/*
   src/kernel/qfontdatabase_win.cpp:extern HFONT stock_sysfont  = 0;
   src/kernel/qfontengine_wce.cpp:static HFONT stock_sysfont  = 0;
   src/kernel/qfontengine_win.cpp:static HFONT stock_sysfont  = 0;
   src/kernel/qpainter_win.cpp:static HFONT  stock_sysfont;
   ...
   $ 

kernel/qfontengine_win.cpp:376:
	warning: assignment to `int' from `double'
kernel/qfontengine_win.cpp:376:
	warning: argument to `int' from `double'
=> This one is just cosmetic but trivial to fix.

kernel/qprinter_win.cpp:340:
	warning: unused variable `QSize margs'
=> This one is just cosmetic but trivial to fix.

[git-p4: depot-paths = "//depot/qt/main/": change = 113309]

+++ src/kernel/qfontengine_win.cpp
@@ -363,7 +363,7 @@ void QFontEngineWin::draw( QPainter *p, int x, int y, const QTextEngine *engine,
 
 	// rotation + scale + translation
 	scale = sqrt( p->m11()*p->m22() - p->m12()*p->m21() );
-	angle = 1800*acos( p->m11()/scale )/M_PI;
+	angle = (int)(1800*acos( p->m11()/scale )/M_PI);
 	if ( p->m12() < 0 )
 	    angle = 3600 - angle;
